% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_facet.R
\name{add_facet}
\alias{add_facet}
\title{Add a facet label to a ggplot object which only contains 1 panel}
\usage{
add_facet(
  p,
  label,
  fill = "#7fc97f",
  color = "black",
  size = 14,
  face = "bold",
  r = 0.25
)
}
\arguments{
\item{p}{the ggplot2 object to be added}

\item{label}{the text to be added as a facet label}

\item{fill}{the strip background color}

\item{color}{the color of the text}

\item{size}{the size of the text, default size is 14}

\item{face}{the font face of the text}

\item{r}{the radius of the rounded corners, a \code{unit} object,
default is unit(0.1, 'snpc').}
}
\value{
a ggplot2 object with a facet label
}
\description{
Add a facet label to a ggplot object which only contains 1 panel
}
\examples{
p <- ggplot(mtcars, aes(disp, drat)) + geom_point()
ggfacet::add_facet(p, label = "test\n (mg/L)\n p = 0.01")
}
\author{
Zhonghui Gai
}
