% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_facet_label.R
\name{edit_facet_label}
\alias{edit_facet_label}
\title{This function is adapted from ggfun package of facet_set function}
\usage{
edit_facet_label(label)
}
\arguments{
\item{label}{the text to ba edited}
}
\value{
a ggplot2 pacakge with modified facet label
}
\description{
This function is adapted from ggfun package of facet_set function
}
\examples{
install.packages("ggfun")
p <- ggplot(mtcars, aes(disp, drat)) + geom_point() + facet_wrap(~am)
p + edit_facet_label(label = c(`0` = "Zero\n p = 0.01\n(mg/L)"))
}
\author{
ZhonghuiGai
}
