% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_facet_border.R
\name{round_facet_border}
\alias{round_facet_border}
\title{Round rectangle border of facet. A swap function of ggfun}
\usage{
round_facet_border(p, r = 0.2)
}
\arguments{
\item{r}{the radius of the rounded corners, a \code{unit} object,
default is unit(0.1, 'snpc').}
}
\value{

}
\description{
Round rectangle border of facet. A swap function of ggfun
}
\examples{
p <- ggplot(mpg, aes(displ, cty)) + geom_point() + facet_grid(cols = vars(cyl))
round_facet_border(p, r = 0.25)
}
\author{
Zhonghui Gai
}
